+++ AspNetExample/Program.cs
@@ -20,7 +20,7 @@ void PrBotInstance_OnLogError(Exception ex, long? id)
     Console.WriteLine(ex.ToString());
 }
 
-void PrBotInstance_OnLogCommon(string msg, Enum typeEvent, ConsoleColor color)
+void PrBotInstance_OnLogCommon(string msg, string typeEvent, ConsoleColor color)
 {
     Console.WriteLine(msg);
 }