+++ PRTelegramBot/Attributes/BaseQueryAttribute.cs
@@ -2,14 +2,33 @@
 
 namespace PRTelegramBot.Attributes
 {
+    /// <summary>
+    /// Базовый атрибут для обработки команд.
+    /// </summary>
+    /// <typeparam name="T">Тип параметра.</typeparam>
     [AttributeUsage(AttributeTargets.Method, Inherited = true)]
     public class BaseQueryAttribute<T> : Attribute, IBotIdentifier, ICommandStore<T>
     {
-        protected List<T> commands = new List<T>();
+        #region Поля и свойства
+
         /// <summary>
-        /// Коллекция команд.
+        /// Команды для методов.
         /// </summary>
-        public IEnumerable<T> Commands { get { return commands.ToList(); } }
+        protected List<T> commands = new List<T>();
+
+        #endregion
+
+        #region ICommandStore
+
+        public IEnumerable<T> Commands
+        {
+            get
+            {
+                return commands.ToList();
+            }
+        }
+
+        #endregion
 
         #region IBotIdentifier
 
@@ -20,9 +39,13 @@ namespace PRTelegramBot.Attributes
 
         #endregion
 
+        #region Конструкторы
+
         public BaseQueryAttribute(long botId = 0)
         {
             BotId = botId;
         }
+
+        #endregion
     }
 }