+++ PRTelegramBot/Attributes/BotHandlerAttribute.cs
@@ -1,5 +1,8 @@
 ﻿namespace PRTelegramBot.Attributes
 {
+    /// <summary>
+    /// Атрибут для работы с DI и экземплярами классов.
+    /// </summary>
     [AttributeUsage(AttributeTargets.Class)]
     public class BotHandlerAttribute : Attribute { }
 }