+++ PRTelegramBot/Attributes/InlineCallbackHandlerAttribute.cs
@@ -1,17 +1,30 @@
 ﻿namespace PRTelegramBot.Attributes
 {
     /// <summary>
-    /// Атрибут для inline команд
+    /// Атрибут для inline команд.
     /// </summary>
     public class InlineCallbackHandlerAttribute<T> : BaseQueryAttribute<Enum> where T : Enum
     {
-        public InlineCallbackHandlerAttribute(params T[] commands) 
-            : this (0, commands) { }
+        #region Конструкторы
 
+        /// <summary>
+        /// Конструктор.
+        /// </summary>
+        /// <param name="commands">Команды.</param>
+        public InlineCallbackHandlerAttribute(params T[] commands)
+            : this(0, commands) { }
+
+        /// <summary>
+        /// Конструктор.
+        /// </summary>
+        /// <param name="botId">Идентификатор бота.</param>
+        /// <param name="commands">Команды.</param>
         public InlineCallbackHandlerAttribute(long botId, params T[] commands) : base(botId)
         {
             foreach (var command in commands)
                 this.commands.Add((Enum)command);
         }
+
+        #endregion
     }
 }