+++ PRTelegramBot/Attributes/InlineCommandAttribute.cs
@@ -1,6 +1,8 @@
 ﻿namespace PRTelegramBot.Attributes
 {
-
+    /// <summary>
+    /// Атрибут для отметки inline команд.
+    /// </summary>
     [AttributeUsage(AttributeTargets.Enum)]
     public class InlineCommandAttribute : Attribute { }
 }