+++ PRTelegramBot/Attributes/ReplyMenuDynamicHandlerAttribute.cs
@@ -2,11 +2,25 @@
 
 namespace PRTelegramBot.Attributes
 {
+    /// <summary>
+    /// Атрибут для обработки dynamic reply методов.
+    /// </summary>
     public class ReplyMenuDynamicHandlerAttribute : BaseQueryAttribute<string>
     {
+        #region Конструкторы
+
+        /// <summary>
+        /// Конструктор.
+        /// </summary>
+        /// <param name="commands">Команды.</param>
         public ReplyMenuDynamicHandlerAttribute(params string[] commands)
             : this(0, commands) { }
 
+        /// <summary>
+        /// Конструктор.
+        /// </summary>
+        /// <param name="botId">Идентификатор бота.</param>
+        /// <param name="commands">Команды.</param>
         public ReplyMenuDynamicHandlerAttribute(long botId, params string[] commands) : base(botId)
         {
             this.commands.AddRange(commands);
@@ -15,9 +29,9 @@ namespace PRTelegramBot.Attributes
                 return;
 
             var dynamicCommand = bot.Options.ReplyDynamicCommands;
-            foreach (var command in commands) 
+            foreach (var command in commands)
             {
-                if(!dynamicCommand.ContainsKey(command))
+                if (!dynamicCommand.ContainsKey(command))
                 {
                     //bot.InvokeErrorLog();
                     continue;
@@ -25,5 +39,7 @@ namespace PRTelegramBot.Attributes
                 this.commands.Add(dynamicCommand[command]);
             }
         }
+
+        #endregion
     }
 }