+++ PRTelegramBot/Attributes/RequireTypeMessageAttribute.cs
@@ -3,18 +3,30 @@
 namespace PRTelegramBot.Attributes
 {
     /// <summary>
-    /// Метод сможет обработать только определенный тип сообщений
+    /// Метод сможет обработать только определенный тип сообщений.
     /// </summary>
     public class RequireTypeMessageAttribute : Attribute
     {
+        #region Поля и свойства
+
         /// <summary>
-        /// Тип сообщения
+        /// Типы сообщений.
         /// </summary>
         public List<MessageType> TypeMessages { get; private set; } = new List<MessageType>();
 
-        public RequireTypeMessageAttribute(params MessageType[] typeData)
+        #endregion
+
+        #region Конструкторы
+
+        /// <summary>
+        /// Конструктор.
+        /// </summary>
+        /// <param name="typeMessages">Тип сообщений.</param>
+        public RequireTypeMessageAttribute(params MessageType[] typeMessages)
         {
-            TypeMessages.AddRange(typeData.ToList());
+            TypeMessages.AddRange(typeMessages.ToList());
         }
+
+        #endregion
     }
 }