+++ PRTelegramBot/Models/PagedResult.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PRTelegramBot.Models
+{
+    /// <summary>
+    /// Струтура класса для постраничного вывода
+    /// </summary>
+    public abstract class PagedResultBase
+    {
+        public int CurrentPage { get; set; }
+        public int PageCount { get; set; }
+        public int PageSize { get; set; }
+        public int RowCount { get; set; }
+
+        public int FirstRowOnPage
+        {
+
+            get { return (CurrentPage - 1) * PageSize + 1; }
+        }
+
+        public int LastRowOnPage
+        {
+            get { return Math.Min(CurrentPage * PageSize, RowCount); }
+        }
+    }
+
+    /// <summary>
+    /// Класс для постраничного вывода с типом лист
+    /// </summary>
+    /// <typeparam name="T">Любой объект</typeparam>
+    public class PagedResult<T> : PagedResultBase where T : class
+    {
+        public IList<T> Results { get; set; }
+
+        public PagedResult()
+        {
+            Results = new List<T>();
+        }
+    }
+}